implement plant rendering

+++ core/src/rendering/o2d/object/position.rs
@@ -4,8 +4,8 @@ use super::layer::LayerO2D;
 pub struct PositionO2D {
     pub tile_x: u8,
     pub tile_y: u8,
-    pub x_offset: u8,
-    pub y_offset: u8,
+    pub x_offset: i8,
+    pub y_offset: i8,
     pub z_index: u8,
 }
 
@@ -28,9 +28,9 @@ impl PositionO2D {
     }
 
     pub fn pixel_position(&self, tile_size: u8) -> (u32, u32) {
-        let x = self.tile_x as u32 * tile_size as u32 + self.x_offset as u32;
-        let y = self.tile_y as u32 * tile_size as u32 + self.y_offset as u32;
-        (x, y)
+        let x = self.tile_x as i32 * tile_size as i32 + self.x_offset as i32;
+        let y = self.tile_y as i32 * tile_size as i32 + self.y_offset as i32;
+        (x as u32, y as u32)
     }
 
     pub fn with_layer(self, layer: impl LayerO2D) -> Self {
@@ -40,7 +40,7 @@ impl PositionO2D {
         }
     }
 
-    pub fn with_offsets(self, x_offset: u8, y_offset: u8) -> Self {
+    pub fn with_offsets(self, x_offset: i8, y_offset: i8) -> Self {
         Self {
             x_offset,
             y_offset,
